refactoring - change script name chillers to chiller_vcc
